fix: fixed `remove()` operation to not create intermediate items
