+++ src/operations/remove/index.js
@@ -1,5 +1,4 @@
-const update = require('../update');
-const { isNumber, extractSubPath } = require('../common');
+const { update, isNumber, extractSubPath } = require('../common');
 
 module.exports = remove;
 
@@ -7,6 +6,10 @@ function remove(object, path) {
     const [ itemPath, valuePath ] = extractSubPath(path);
 
     return update(object, itemPath, (item) => {
+        if (!item) {
+            return item;
+        }
+
         if (Array.isArray(item) && isNumber(valuePath)) {
             return item
                 .slice(0, valuePath)
@@ -19,5 +22,5 @@ function remove(object, path) {
         } = item;
 
         return newItem;
-    });
+    }, false);
 }
